+++ src/hcom/transport/shm/net_shm_driver_oob.cpp
@@ -242,7 +242,7 @@ NResult NetDriverShmWithOOB::CreateWorkers()
     NResult result = NN_OK;
 
     std::vector<uint16_t> workerGroups;
-    std::vector<std::pair<uint8_t, uint8_t>> workerGroupCpus;
+    std::vector<std::pair<uint32_t, uint32_t>> workerGroupCpus;
     std::vector<int16_t> flatWorkerCpus;
     std::vector<int16_t> workerThreadPriority;
